ARROW-13544 [Java]: Remove APIs that have been deprecated for long (Changes to JDBC)
See https://issues.apache.org/jira/browse/ARROW-13544

According to the discussion in #10864 (comment), we want to split the task into multiple parts.

This PR is for the changes related to the JDBC adapter

Closes #10912 from liyafan82/fly_0811_depj

Authored-by: liyafan82 <fan_li_ya@foxmail.com>
